fix similarity index, up 1.1.7

+++ lib/finnhub_ruby/models/company_profile.rb
@@ -166,7 +166,7 @@ module FinnhubRuby
         :'ipo' => :'Date',
         :'market_capitalization' => :'Float',
         :'share_outstanding' => :'Float',
-        :'employee_total' => :'Integer',
+        :'employee_total' => :'Float',
         :'logo' => :'String',
         :'finnhub_industry' => :'String'
       }
